requirements.txt
cwiid==0.6.0
pyOSC==0.3.5b5294
